pyproject.toml
[tool.poetry]
name = "skymarket"
version = "0.1.0"
description = ""
authors = ["medved"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-rest = "^0.8.7"
django-rest-framework = "^0.1.0"
pillow = "^9.4.0"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
djoser = "^2.1.0"
django-cors-headers = "^3.13.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.0.0"}
drf-spectacular = "^0.24.2"
drf-nested-routers = "^0.93.4"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
